+++ src/ui/buttons/button.jsx
@@ -44,8 +44,9 @@ type IndividualButtonProps = {|
 
 export function Button({ fundingSource, style, multiple, locale, env, fundingEligibility, i, nonce, flow, vault,
     userIDToken, personalization, onClick = noop, content, tagline, commit, experiment, instrument, experience } : IndividualButtonProps) : ElementNode {
-
-    const inlineExperience = experience === EXPERIENCE.INLINE;
+    
+    const { custom, layout, shape } = style;
+    const inlineExperience = experience === EXPERIENCE.INLINE && custom;
     const fundingConfig = getFundingConfig()[fundingSource];
 
     if (!fundingConfig) {
@@ -95,8 +96,6 @@ export function Button({ fundingSource, style, multiple, locale, env, fundingEli
             preventClickFocus(el);
         }
     };
-
-    const { custom, layout, shape } = style;
     
     const labelText = typeof fundingConfig.labelText === 'function' ?  fundingConfig.labelText({ content, fundingEligibility }) : (fundingConfig.labelText || fundingSource);