+++ src/zoid/buttons/util.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { supportsPopups as userAgentSupportsPopups, isAndroid, isChrome, isIos, isSafari, isSFVC, type Experiment, isDevice, isTablet, getElement, isLocalStorageEnabled } from '@krakenjs/belter/src';
-import { ENV, FUNDING } from '@paypal/sdk-constants/src';
-import { getEnableFunding, getDisableFunding, createExperiment, getFundingEligibility, getPlatform, getComponents, getEnv } from '@paypal/sdk-client/src';
+import { COUNTRY, CURRENCY, ENV, FPTI_KEY, FUNDING, type LocaleType } from '@paypal/sdk-constants/src';
+import { getEnableFunding, getDisableFunding, getLogger, createExperiment, getFundingEligibility, getPlatform, getComponents, getEnv, type FundingEligibilityType } from '@paypal/sdk-client/src';
 import { getRefinedFundingEligibility } from '@paypal/funding-components/src';
 
 import type { Experiment as EligibilityExperiment } from '../../types';
@@ -295,4 +295,41 @@ export function getButtonSize (props : ButtonProps, container : string | HTMLEle
             return BUTTON_SIZE.HUGE;
         }
     }
+}
+
+type InlineCheckoutEligibilityProps = {|
+    commit : boolean,
+    createBillingAgreement? : Function,
+    currency : string,
+    disableFunding : $ReadOnlyArray<$Values<typeof FUNDING>>,
+    fundingEligibility : FundingEligibilityType,
+    layout : $Values<typeof BUTTON_LAYOUT>,
+    locale : LocaleType,
+    merchantID? : $ReadOnlyArray<string>,
+    vault : boolean
+|};
+
+export function isInlineXOEligible({ props } : {| props : InlineCheckoutEligibilityProps |}) : boolean {
+    const { commit, currency, createBillingAgreement, disableFunding, fundingEligibility, layout, locale, merchantID, vault } = props;
+
+    const isEligible = (
+        locale.country === COUNTRY.US &&
+        commit === true &&
+        !createBillingAgreement &&
+        currency === CURRENCY.USD &&
+        (disableFunding?.indexOf(FUNDING.CARD) === -1) &&
+        (fundingEligibility?.card?.eligible || false) &&
+        layout === BUTTON_LAYOUT.VERTICAL &&
+        merchantID?.length === 0 &&
+        vault === false
+    );
+
+    getLogger()
+        .info('isInlineXOEligible props', { props: JSON.stringify(props) })
+        .info('isInlineXOEligible eligible', { eligible: String(isEligible) })
+        .track({
+            [ FPTI_KEY.TRANSITION ]: `inline_xo_eligibility_${ String(isEligible) }`
+        }).flush();
+
+    return isEligible;
 }